+++ aiarena/frontend/views/__init__.py
@@ -1,0 +1,34 @@
+# ruff: noqa: F401
+from .arena_client_detail import ArenaClientView
+from .arena_client_list import ArenaClients
+from .author_detail import AuthorDetail
+from .author_list import AuthorList
+from .bot_competition_stats_detail import BotCompetitionStatsDetail
+from .bot_competition_stats_elo_graph_update_plot import BotCompetitionStatsEloGraphUpdatePlot
+from .bot_data_download import BotDataDownloadView
+from .bot_detail import BotDetail
+from .bot_list import BotDownloadableList, BotList
+from .bot_update import BotUpdate
+from .bot_zip_download import BotZipDownloadView
+from .competition_detail import CompetitionDetail
+from .competition_list import CompetitionList
+from .competition_participation_list import CompetitionParticipationList
+from .competition_participation_update import CompetitionParticipationUpdate
+from .health_check import health_check, health_check_with_db
+from .http_500 import http_500
+from .index import Index
+from .match_detail import MatchDetail, MatchTagFormView
+from .match_log_download import MatchLogDownloadView
+from .match_queue import MatchQueue
+from .project_finance import project_finance
+from .recent_results import RecentResults
+from .request_match import RequestMatch
+from .round_detail import RoundDetail
+from .user_profile import (
+    BotUpload,
+    UnlinkDiscordView,
+    UnlinkPatreonView,
+    UserProfile,
+    UserProfileUpdate,
+    UserTokenDetailView,
+)
